server: fixed permissions func

+++ server/api/messages.js
@@ -2,7 +2,7 @@ const database = require('../database/messages');
 const rooms = require('../database/rooms');
 const users = require('../database/users');
 
-async function roomPermissions(req, res, callback){
+async function roomPermissions(req, res){
 
     try {
         if(!req.params.room){
@@ -23,11 +23,10 @@ async function roomPermissions(req, res, callback){
         let rooms_id = user.rooms;
         
         if(rooms_id.indexOf(req.params.room) != -1){
-            callback(room);
+            return room;
         } else {
-            return res.status(400).json({ status: 400, message: 'You are not a member of this room'}); 
+            return false;
         }
-
 
     } catch (error) {
         console.log(error);
@@ -75,11 +74,16 @@ async function addMessage(req, res){
         if(!req.body.message){
             return res.status(400).json({ status: 400, message: 'Message is required'}); 
         }
-        roomPermissions(req, res, (room) =>{
-            let users = room.users;
-            await database.addMessage(req.session.userid, req.body.message, req.params.room);   
-            res.status(200).json({status: 200, message: "success"});
-        })        
+        let permissions = roomPermissions(req, res);
+
+        if (!permissions){
+            return res.status(400).json({ status: 400, message: 'You are not a member of this room'}); 
+        }
+
+        let users = permissions.users;
+        await database.addMessage(req.session.userid, req.body.message, req.params.room);   
+        res.status(200).json({status: 200, message: "success"});
+
     } catch (error) {
         console.log(error);
         return res.status(500).json({ status: 500, message: error});
@@ -92,10 +96,15 @@ async function readMessage(req, res){
             return res.status(400).json({ status: 400, message: 'Receiver is required'}); 
         }
 
-        roomPermissions(req, res, () => {
-            await database.readMessage(req.params.message_id, req.params.room);
-            res.status(200).json({status: 200, message: "success"});
-        })
+        let permissions = roomPermissions(req, res);
+        
+        if (!permissions){
+            return res.status(400).json({ status: 400, message: 'You are not a member of this room'}); 
+        }
+
+        await database.readMessage(req.params.message_id, req.params.room);
+        res.status(200).json({status: 200, message: "success"});
+
     } catch (error) {
         console.log(error);
         return res.status(500).json({ status: 500, message: error});
@@ -108,11 +117,16 @@ async function deleteMessage(req, res){
         if(!req.params.message_id){
             return res.status(400).json({ status: 400, message: 'Receiver is required'}); 
         }
+        
+        let permissions = roomPermissions(req, res);
+        
+        if (!permissions){
+            return res.status(400).json({ status: 400, message: 'You are not a member of this room'}); 
+        }
 
-        roomPermissions(req, res, () => {
-            await database.deleteMessage(req.session.userid, req.params.message_id, req.params.room);
-            res.status(200).json({status: 200, message: "success"});
-        })
+        await database.deleteMessage(req.session.userid, req.params.message_id, req.params.room);
+        res.status(200).json({status: 200, message: "success"});
+
     } catch (error) {
         console.log(error);
         return res.status(500).json({ status: 500, message: error});
@@ -125,10 +139,15 @@ async function hideMessage(){
             return res.status(400).json({ status: 400, message: 'Receiver is required'}); 
         }
 
-        roomPermissions(req, res, () => {
-            await database.hideMessage(req.session.userid, req.params.message_id, req.params.room);
-            res.status(200).json({status: 200, message: "success"});
-        })
+        let permissions = roomPermissions(req, res);
+        
+        if (!permissions){
+            return res.status(400).json({ status: 400, message: 'You are not a member of this room'}); 
+        }
+
+        await database.hideMessage(req.session.userid, req.params.message_id, req.params.room);
+        res.status(200).json({status: 200, message: "success"});
+        
     } catch (error) {
         console.log(error);
         return res.status(500).json({ status: 500, message: error});
@@ -141,17 +160,22 @@ async function getMessage(req, res){
             return res.status(400).json({ status: 400, message: 'Receiver is required'}); 
         }
 
-        roomPermissions(req, res, () => {
-            let message = database.getMessage(req.params.message_id, req.params.room);
-            let user = users.getUser(message.sender);
-            res.status(200).json({
-                sender_id: message.sender,
-                sender_nickname: user.nickname,
-                sender_avatar: user.avatar,
-                message: message.message,
-                status: message.status
-            });
-        })
+        let permissions = roomPermissions(req, res);
+        
+        if (!permissions){
+            return res.status(400).json({ status: 400, message: 'You are not a member of this room'}); 
+        }
+
+        let message = database.getMessage(req.params.message_id, req.params.room);
+        let user = users.getUser(message.sender);
+        res.status(200).json({
+            sender_id: message.sender,
+            sender_nickname: user.nickname,
+            sender_avatar: user.avatar,
+            message: message.message,
+            status: message.status
+        });
+        
     } catch (error) {
         console.log(error);
         return res.status(500).json({ status: 500, message: error});
@@ -160,26 +184,31 @@ async function getMessage(req, res){
 
 async function getMessages(req, res){
     try {
-        roomPermissions(req, res, () => {
-            let offset, limit, data = [];
-            let users = userData(req, res);
-            offset = req.query.offset ? Number(req.query.offset) : 0;
-            limit = req.query.limit ? Number(req.query.limit) : 20;
-            let messages = await database.getMessages(req.session.userid, 
-                                                      req.params.room, 
-                                                      offset, 
-                                                      limit);
-            for(message of messages){
-                data.push({
-                    message: message.message,
-                    sender_id: message.sender,
-                    sender_nickname: users[message.sender].nickname,
-                    sender_avatar: users[message.sender].avatar,
-                    status: message.status
-                })
-            }
-            res.status(200).json(data);
-        })
+        let permissions = roomPermissions(req, res);
+        
+        if (!permissions){
+            return res.status(400).json({ status: 400, message: 'You are not a member of this room'}); 
+        }
+        
+        let offset, limit, data = [];
+        let users = userData(req, res);
+        offset = req.query.offset ? Number(req.query.offset) : 0;
+        limit = req.query.limit ? Number(req.query.limit) : 20;
+        let messages = await database.getMessages(req.session.userid, 
+                                                    req.params.room, 
+                                                    offset, 
+                                                    limit);
+        for(message of messages){
+            data.push({
+                message: message.message,
+                sender_id: message.sender,
+                sender_nickname: users[message.sender].nickname,
+                sender_avatar: users[message.sender].avatar,
+                status: message.status
+            })
+        }
+        res.status(200).json(data);
+        
     } catch (error) {
         console.log(error);
         return res.status(500).json({ status: 500, message: error});
